UPSTREAM: cbfstool: Remove location pointer from parse_elf_to_stage()
The *location argument to parse_elf_to_stage() is a relic from code all
the way back to 2009 where this function was still used to parse XIP
stages. Nowadays we have a separate parse_elf_to_xip_stage() for that,
so there is no need to heed XIP concerns here. Having a pointer to
represent the location in flash is absolutely irrelevant to a non-XIP
stage, and it is used incorrectly -- we just get lucky that no code path
in cbfstool can currently lead to that value being anything other than
0, otherwise the adjustment of data_start to be no lower than *location
could easily screw things up. This patch removes it.

BUG=none
BRANCH=none
TEST=none

Change-Id: I7fb57f2180d511b2f95a5e19b29b7d60984c1abf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ff61a39e90e0f8cb94a3dd015a2ebd6cae728cc6
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ia7f850c0edd7536ed3bef643efaae7271599313d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49369
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2704320
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>