UPSTREAM: Makefile.inc: Avoid --emit-relocs on RISC-V
There seems to be a bug[1] in the GNU linker for the RISC-V architecture
triggered by symbols that are more than 2GB offset from the program
counter. My next patch is introducing symbols like that and stuck on
this problem. The code path that runs into the issue is only taken when
passing the --emit-relocs flag, which is really only needed for building
rmodules. Since RISC-V platforms don't use any rmodules at the moment,
let's disable the flag on RISC-V until the issue can be fixed in the
toolchain.

[1]: https://sourceware.org/bugzilla/show_bug.cgi?id=27180

BUG=none
BRANCH=none
TEST=none

Change-Id: Iea44da784ffe60efeb757854cf8e2d3c5759b17d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7d4ef16efebe489b247b55740bc284b36d6879e2
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I784a506034325c0ba937589416acaafbf80080e2
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49449
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2644983
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>