UPSTREAM: trogdor: Initialize BACKLIGHT_ENABLE to 0, only turn it on in payload
The BACKLIGHT_ENABLE pin on this board unfortunately defaults to a
pull-up on power on, meaning the backlight is immediately enabled. Best
we can do about that is to turn it off again early and wait until it is
actually correct in the panel power sequence to turn it back on.

Some panels want an explicit 80ms delay after training the eDP
connection before the backlight is turned on (this is probably just to
avoid temporary display artifacts, but whatever). We don't want to
busy-wait that extra time, so instead just delegate turning on that GPIO
to the payload (which is also in charge of the backlight PWM already).

BRANCH=trogdor
BUG=b:177667751
TEST=Booted Lazor

Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Id8dafbdcb40175fbc9205276eee698583b971873
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49495
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Douglas Anderson <dianders@chromium.org>

Cq-Depend: chromium:2634286
Change-Id: Ic61beb270ef97c17bf930fb082901b3edeb9372a
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2638421
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>