UPSTREAM: soc/amd/common/acpi: Add _UID for PNP0C02 devices
When MAINBOARD_HAS_SPEAKER is false, the SPKR gets _HID PNP0C02. This
conflicts with the LDRC device. PNP0C02 is also used other places in the
picasso code base, so I chose a random _UID for each device. The _UIDs
are unique in the code base so it's easy to search for duplicates.

BUG=b:175146875
TEST=Boot trembyle to linux

Change-Id: I59d1ad6dde75ed41d20c029590e19780dfc64967
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a626d2ed6f15748f40500cef1ec074ed9fd785df
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: I01be41515e011293e90a6b42b8e34de8ec3ffc18
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49813
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2647392
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>