UPSTREAM: trogdor: Explicitly initialize display pins in bootblock
This patch adds explicit initializations for the remaining named display
(power) control GPIOs to the bootblock GPIO init code. These pins are
usually mapped to pins that are already configured to pull-downs on
power-on reset so this wasn't really required, but we have already moved
them around so often that you never know when EEs might one day move
them to a pin with a different power-on reset configuration, so it's
better to be explicit.

In one particular case, GPIO(67) (used by CoachZ rev1+ but not by
anything else for the EN_PP3300_DX_EDP pin) is not actually a pull-down
on boot, even though that is claimed by the datasheet. This is likely
due to the fact that it can serve as the SPI_HOLD pin for the boot flash
QSPI bus, so even though our board's boot flash doesn't really use that
pin, it seems that the boot ROM still configures it as such.

BUG=none
BRANCH=none
TEST=none

Change-Id: I333596ab3cb1a5084716120a87a16168007bb4b7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ac80610b510e5e1d270b711bbc9632e3ab30f5de
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I533baa962d2dfc87cfa510f442ed2e8912e0e5b0
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49810
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: mturney mturney <mturney@codeaurora.org>
Original-Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2647395
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>