UPSTREAM: soc/amd/picasso/acpi: Fix PCI0 MMIO window
The PCI0 MMIO window was defined between TOM and 4 GiB. This was
overlapping with the FCH MMIO devices. The first MMIO device after TOM
is the FCH IOAPIC.

This wasn't causing a problem for linux other than the fact that
/proc/iomem showed all the MMIO devices under the PCI root bridge.
On Windows this was causing all the MMIO devices to have conflicting
resource errors.

BUG=b:175146875
BRANCH=zork
TEST=Boot linux and verify peripherals all work. Boot windows and
verify the i2c controllers show up. The GPIO controller still has a
problem related to power.

Change-Id: Ic40290c5e2c471829be7fee97ec587c649721543
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c3d7846a336e23168dc159e09e7c6b412c7a9cb7
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: Idc409f1318e6da5a693ccbb3da74aafd13f1e058
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/49853
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2655444
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>