UPSTREAM: sb,soc/intel: Add wake source fields in GNVS
For the moment, these are most not used but become a necessity
for a unified <soc/nvs.h> approach.

They would be required for the implementation of _SWS method
for OSPM to determine the reason for system waking up. The related
hardware registers are present with these platforms.

It's expected that ACPI power-management related GNVS entries are
grouped together to form a single struct in later works.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic5319f0e534bec615ca4fbd5dcde3f1bf0bb9cd6
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8fee9951d30d03b4bca16c198b887c5415418c12
Original-Change-Id: I6d31d39ac1017cd6fdf0ac66b418d1fbb1edf8e0
Original-Signed-off-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50193
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2679033
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>