UPSTREAM: soc/amd/picasso/psp: move soc_get_mbox_address to common PSP gen2 code
The function to get the PSP mailbox address is the same on Picasso and
Cezanne, so move it to the common PSP generation 2 code. The function is
only used in the same compilation unit, but it can't be marked as static
due to the function prototype in amdblocks/psp.h that is still needed
for Stoneyridge.

BUG=none
BRANCH=none
TEST=none

Change-Id: I60464193169653e34dfcefaa2911570c901c43ad
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 84439c26d8afd0079ce9cbaddf11c911974c3b1b
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ieea91ef76523d303f948d29ef48e3b2e56293f26
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50151
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2663269
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>