UPSTREAM: drivers/intel/fsp2_0/memory_init: check if UPD struct has expected size
If the UPD size in coreboot sizes mismatches the one from the FSP-M
binary, we're running into trouble. If the expected size is smaller than
the UPD size the FSP provides, call die(), since the target buffer isn't
large enough so only the beginning of the UPD defaults from the FSP will
get copied into the buffer. We ran into the issue in soc/amd/cezanne,
where the UPD struct in coreboot was smaller than the one in the FSP, so
the defaults didn't get completely copied.

TEST=Mandolin still boots.

Change-Id: Icd1ed7b71996099d424b110b1e74dbc193da57e5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a09559501c06e3cab420908bd667af4486ec0ba5
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ia7e9f6f20d0091bbb4abfd42abb40b485da2079d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50241
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2714582
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>