UPSTREAM: soc/amd/cezanne/fch: add ACPI I/O port setup
The offsets of ACPI_CPU_CONTROL and ACPI_GPE0_BLK match the ones from
the reference code, but not the PPR. I've submitted a change request for
the PPR, so this mismatch might go away in the future. The case for
HAVE_SMI_HANDLER will be implemented in a future patch. If that one ends
up being identical to the function in soc/amd/picasso, I'll move it to
the common AMD SoC code.

BUG=none
BRANCH=none
TEST=none

Change-Id: I72f00fc3889ac49a1b6a7ca782194a705be3a407
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7e703d77b2cb2189863d2a2e87f7de2913e20307
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: If80b841df12d351d5a0c1e0d2e7bf1e31b03447f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50270
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2679030
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>