UPSTREAM: spd_bin: Replace get_spd_cbfs_rdev() with spd_cbfs_map()
In pursuit of the goal of eliminating the proliferation of raw region
devices to represent CBFS files outside of the CBFS core code, this
patch removes the get_spd_cbfs_rdev() API and instead replaces it with
spd_cbfs_map() which will find and map the SPD file in one go and return
a pointer to the relevant section. (This makes it impossible to unmap
the mapping again, which all but one of the users didn't bother to do
anyway since the API is only used on platforms with memory-mapped
flash. Presumably this will stay that way in the future so this is not
something worth worrying about.)

BUG=none
BRANCH=none
TEST=none

Change-Id: I22760609ace201dc265e4347f5005dbf512b146d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a9b44f4c79078210fe9966daf2412cc222c2d0a9
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Iec7571bec809f2f0712e7a97b4c853b8b40702d1
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50350
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Wim Vervoorn <wvervoorn@eltan.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2766754
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>