UPSTREAM: trogdor: Add fingerprint power sequencing
Some Trogdor variants will include a fingerprint sensor, so this patch
adds support for its power sequencing. There is a requirement that the
fingerprint power needs to be *off* for at least 200ms, and when it is
turned back on it needs to stabilize for at least 3.5ms before taking
the FPMCU out of reset. We meet these timing requirements by splitting
the sequence across bootblock, romstage and ramstage. On current Trogdor
boards we measured <end of bootblock> to <end of romstage> at ~430ms and
<end of romstage> to <start of ramstage> at 12ms, so we easily meet the
required numbers this way.

BRANCH=trogdor
BUG=b:170284663

Change-Id: Iacbbdc7ae2e1ad217bff397dd196e680b8c68201
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a39c5772572dff6f324ef5437d36223afa0f7f69
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Iccd77e6e1c378110fca2b2b7ff1f534fce54f8ea
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50504
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Shelley Chen <shchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2695676
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>