UPSTREAM: soc/amd: add and use fch_enable_hpet_decode
On Picasso we missed setting this bit in coreboot and since the default
after reset is 0, we had to rely on the FSP to set this bit. Stoneyridge
and Cezanne have the HPET decode enable bit in the same position in the
same register. In the ACPI table entry written by
southbridge_write_acpi_tables the HPET entry gets added, so we should
make sure that we enable the decode.

TEST=HPET still works on Mandolin.

Change-Id: If634ff6b2d9299b57913b38a858316b107d4579c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 34fc29ae96d2ae6e0d021c84fc3ff8d56ec38898
Original-Change-Id: Ie98dae1d6036748f700f884d4b9653f2e59c24da
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50512
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2695674
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>