UPSTREAM: soc/amd/picasso: introduce and use chipset device tree
The chipset devicetree only has the essential PCIe devices enabled that
are needed for the SoC code to work. It also defines aliases for all
PCIe devices that can be used to reference the devices in the mainboard-
specific devicetrees and devicetree overrides. To make the change easier
to review that part will be done in a follow-up patch.

Despite missing in the PPR, device pci 18.7 exists on Picasso.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2d9acc1399baa3cca2f5253b3bd23f1e27aaac49
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c4eb45fa85d9860ce94829c6c977b9e28a297bf9
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I6b7c3fd32579a23539594672593a243172c161c7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50626
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2932504
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>