UPSTREAM: cpu/x86/smm.h: Remove smm runtime pointer from smm_loader_params
struct smm_loader_params is a struct that is passed around in the
ramstage code to set up either the relocation handler or the permanent
handler. At the moment no parameters in the stub 'smm_runtime' are
referenced so it can be dropped. The purpose is to drop the
smm_runtime struct from the stub as it is already located in the
permanent handler.

BUG=none
BRANCH=none
TEST=none

Change-Id: I14d33ab1bc401ad504e2cda6b0fd1bcf0077b0be
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6f9b1a904923cab9905bd6521738e7b6617cf3c5
Original-Change-Id: I09c1b649b5991f55b5ccf57f22e4a3ad4c9e4f03
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50766
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2775639
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>