UPSTREAM: util/ifittool: Add an option to set the FIT pointer a CBFS file
The purpose of this is to eventually move the FIT table out of the
bootblock, generate it separately as a cbfs file and then have the FIT
pointer point to that cbfs file.

TESTED: extracted a FIT table using dd, added it as a cbfs file and see
that the FIT pointer correctly points to it. Also test that trying to
add a non valid FIT cbfs file results in an error.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6a5f4987b32d1f5c3b50bdec4aa8d9c68d58bd0d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e9e4e54e27d7d6ff987f694d183534a7f1713c04
Original-Change-Id: I6e38b7df31e6b30f75b0ae57a5332f386e00f16b
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50925
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Reviewed-by: Christian Walter <christian.walter@9elements.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2775634
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>